requirements.txt
aiogram==2.25.1
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
asyncpg==0.27.0
attrs==23.1.0
Babel==2.9.1
certifi==2023.7.22
charset-normalizer==3.3.1
frozenlist==1.4.0
greenlet==3.0.1
idna==3.4
magic-filter==1.0.12
multidict==6.0.4
python-dotenv==1.0.0
pytz==2023.3.post1
SQLAlchemy==2.0.8
typing_extensions==4.8.0
yarl==1.9.2
